added IsKnownAtCompileTime() in multi-index transform and tensor descriptor
