+++ composable_kernel/include/tensor_description/dynamic_tensor_descriptor.hpp
@@ -201,6 +201,19 @@ struct DynamicTensorDescriptor
         return VisibleDimensionIds{};
     }
 
+    __host__ __device__ static constexpr bool IsKnownAtCompileTime()
+    {
+        bool is_known = true;
+
+        static_for<0, Transforms::Size(), 1>{}([&](auto i) {
+            is_known &=
+                remove_cv_t<remove_reference_t<decltype(Transforms{}[i])>>::IsKnownAtCompileTime();
+        });
+
+        return is_known && is_known_at_compile_time<ElementSize>::value &&
+               is_known_at_compile_time<ElementSpaceSize>::value;
+    }
+
     __host__ __device__ void Print() const
     {
         printf("{");
